Properly handle completedDate for Revolut
Report uses completedDate to determine whether or not a transaction should be included in the report. Revolut data now includes not just the date but also the time. While this is useful, it alters the report's decision which results in some transactions being incorrectly included. To match AIB transaction data, the time is now stripped from Revolut transaction data when being handled by the report.